# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Ticketing API",
        "description": "API untuk mengelola berbagai resource",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/statuses": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Get all statuses",
                "operationId": "e96ea40c6f497546795df9ec612e1928",
                "responses": {
                    "200": {
                        "description": "List of statuses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Status"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Create a new status",
                "operationId": "85739a0412fc324e86e6b45af20c7f0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Status"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Status created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statuses/{id}": {
            "put": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Update status",
                "operationId": "7ba406b9e828f2ce3a0583d7c56583f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Status"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Delete status",
                "operationId": "4bcdd7d14d0707a62e62d2315c6149e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Status deleted successfully"
                    },
                    "404": {
                        "description": "Status not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Status": {
                "required": [
                    "status_name",
                    "color"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status_name": {
                        "type": "string",
                        "example": "Pending"
                    },
                    "color": {
                        "type": "string",
                        "example": "#FF0000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-17T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-17T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}